Reduced setting option complexity
